Redisign of register
