+++ sync-snack/app/components/register/RegisterComponent.tsx
@@ -1,47 +1,24 @@
 "use client";
 import React, { useState } from 'react';
-import { Box, useToast, Button, Flex, Step, StepIcon, StepIndicator, StepNumber, StepSeparator, StepStatus, StepTitle, Stepper } from '@chakra-ui/react';
+import { Box, useToast, Button, FormControl, FormLabel, Input, Flex } from '@chakra-ui/react';
 import { useRouter } from 'next/navigation';
-import AccountDetails from './step-one/AccountDetails';
-import UserProfile from './step-two/UserProfile';
-import GroupInformation from './step-three/GroupInformation';
-
-const steps = [
-  { title: 'Account Details', component: AccountDetails },
-  { title: 'User Profile', component: UserProfile },
-  { title: 'Group Information', component: GroupInformation },
-];
 
 export default function RegisterComponent() {
   const [isLoading, setIsLoading] = useState(false);
   const toast = useToast();
   const router = useRouter();
-  const [activeStep, setActiveStep] = useState(0);
   const [formData, setFormData] = useState({
     email: '',
     password: '',
     confirmPassword: '',
-    firstName: '',
-    lastName: '',
-    groupChoice: '',
-    groupName: '',
-    groupPassword: '',
-    groupDescription: '',
   });
 
-  const handleNext = () => setActiveStep((prev) => prev + 1);
-  const handleBack = () => setActiveStep((prev) => prev - 1);
-
   const handleInputChange = (e: any) => {
-    if (e.target) {
-      const { name, value } = e.target;
-      setFormData({ ...formData, [name]: value });
-    } else {
-      setFormData({ ...formData, groupChoice: e });
-    }
+    const { name, value } = e.target;
+    setFormData({ ...formData, [name]: value });
   };
 
-  const showToast = (title: string, description: string, status: 'success' | 'error' | 'warning' | 'info') => {
+  const showToast = (title: string, description: string, status: 'success' | 'error' | 'info' | 'warning') => {
     toast({
       title,
       description,
@@ -57,21 +34,9 @@ export default function RegisterComponent() {
 
     try {
       const isEmailValid = await isUserEmailValid();
-
+      console.log(isEmailValid)
       if (isEmailValid) {
-        showToast('Error', "That user already exists. Can't register this account", 'error');
-        return;
-      }
-
-      let groupId;
-      if (formData.groupChoice === 'create') {
-        groupId = await createGroup();
-      } else {
-        groupId = await joinGroup();
-      }
-
-      if (!groupId) {
-        showToast('Error', 'Failed to create or join group', 'error');
+        showToast('Error', "Your email is not valid", 'error');
         return;
       }
 
@@ -82,55 +47,20 @@ export default function RegisterComponent() {
         return;
       }
 
-      await createUserProfile(userData.userId, groupId);
-
-      showToast('Success', 'Registration successful!', 'success');
+      showToast('Success', 'We have sent you an email verification. Please verify your email.', 'success');
       setTimeout(() => {
-        router.push('/login');
-      }, 2000);  // Redirect after 2 seconds
-    } catch (error) {
-      console.error('Registration error:', error);
-      showToast('Error', 'An unexpected error occurred. Please try again.', 'error');
+        //router.push('https://mail.google.com/mail/u/0/#inbox');
+      }, 2000);
+    } catch (error: any) {
+      if (error.message === 'EmailAlreadyInUse') {
+        showToast('Error', 'This email is already in use.', 'error');
+      } else {
+        console.error('Registration error:', error);
+        showToast('Error', 'An unexpected error occurred. Please try again.', 'error');
+      }
     } finally {
       setIsLoading(false);
     }
-  };
-
-  const createGroup = async () => {
-    const response = await fetch(`http://localhost:8080/api/groups/create`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({
-        name: formData.groupName,
-        password: formData.groupPassword,
-        description: formData.groupDescription,
-      }),
-    });
-
-    if (!response.ok) {
-      throw new Error('Failed to create group');
-    }
-
-    const groupData = await response.json();
-    return groupData.groupId;
-  };
-
-  const joinGroup = async (): Promise<string> => {
-    const response = await fetch(`http://localhost:8080/api/groups/join`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({
-        name: formData.groupName,
-        password: formData.groupPassword,
-      }),
-    });
-
-    if (!response.ok) {
-      throw new Error('Failed to join group');
-    }
-
-    const groupData = await response.json();
-    return groupData.groupId;
   };
 
   const registerUser = async () => {
@@ -143,119 +73,78 @@ export default function RegisterComponent() {
       }),
     });
 
+    if (response.status === 400) {
+      throw new Error('EmailAlreadyInUse');
+    }
+
     if (!response.ok) {
       throw new Error('Failed to register user');
     }
 
+
     return response.json();
   };
 
-  const createUserProfile = async (userId: string, groupId: string) => {
-    const userProfileData = new FormData();
-    const jsonBlob = new Blob([JSON.stringify({
-      userId,
-      groupId,
-      firstName: formData.firstName,
-      lastName: formData.lastName,
-    })], { type: 'application/json' });
-
-    userProfileData.append('body', jsonBlob, 'body.json');
-
-    const response = await fetch(`http://localhost:8080/api/profiles/create`, {
-      method: 'POST',
-      body: userProfileData,
-    });
-
-    if (!response.ok) {
-      throw new Error('Failed to create user profile');
+  const isUserEmailValid = async () => {
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+  
+    if (!emailRegex.test(formData.email)) {
+      return true;
     }
-  };
-
-  const isUserEmailValid = async (): Promise<boolean> => {
+  
     try {
       const response = await fetch('http://localhost:8080/api/users/check', {
         method: 'GET',
-        body: formData.email
+        body: formData.email,
       });
+  
       return response.ok;
     } catch {
       return false;
     }
   };
-
-  const isStepComplete = (step: number) => {
-    const isValidEmail = (email: string) => {
-      const emailRegex = /@syncsnack/i;
-      return emailRegex.test(email);
-    };
-
-    switch (step) {
-      case 0:
-        return (
-          formData.email &&
-          isValidEmail(formData.email) &&
-          formData.password &&
-          formData.confirmPassword &&
-          (formData.password === formData.confirmPassword)
-        );
-      case 1:
-        return formData.firstName && formData.lastName;
-      case 2:
-        if (formData.groupChoice === 'join') {
-          return formData.groupName && formData.groupPassword;
-        } else if (formData.groupChoice === 'create') {
-          return formData.groupName && formData.groupDescription && formData.groupPassword;
-        }
-        return false;
-      default:
-        return false;
-    }
-  };
-
-  const CurrentStepComponent = steps[activeStep].component;
+  
 
   return (
-    <Box className="max-w-2xl mx-auto mt-10 p-6 bg-white rounded-lg shadow-md">
-      <Stepper index={activeStep} className="mb-8">
-        {steps.map((step, index) => (
-          <Step key={index}>
-            <StepIndicator>
-              <StepStatus
-                complete={<StepIcon />}
-                incomplete={<StepNumber />}
-                active={<StepNumber />}
-              />
-            </StepIndicator>
-            <Box flexShrink='0'>
-              <StepTitle>{step.title}</StepTitle>
-            </Box>
-            <StepSeparator />
-          </Step>
-        ))}
-      </Stepper>
-      <CurrentStepComponent formData={formData} handleInputChange={handleInputChange} />
-      <Flex justifyContent="space-between" mt={6}>
-        <Button onClick={handleBack} isDisabled={activeStep === 0}>
-          Back
+    <Box className="max-w-md mx-auto mt-10 p-6 bg-white rounded-lg shadow-md">
+      <FormControl className="mb-4">
+        <FormLabel>Email</FormLabel>
+        <Input
+          name="email"
+          value={formData.email}
+          onChange={handleInputChange}
+          type="email"
+        />
+      </FormControl>
+      <FormControl className="mb-4">
+        <FormLabel>Password</FormLabel>
+        <Input
+          name="password"
+          value={formData.password}
+          onChange={handleInputChange}
+          type="password"
+        />
+      </FormControl>
+      <FormControl className="mb-4">
+        <FormLabel>Confirm Password</FormLabel>
+        <Input
+          name="confirmPassword"
+          value={formData.confirmPassword}
+          onChange={handleInputChange}
+          type="password"
+        />
+      </FormControl>
+      <Flex justifyContent="center" mt={6}>
+        <Button
+          colorScheme="orange"
+          onClick={handleSubmit}
+          isDisabled={isLoading || !(formData.email && formData.password && formData.password === formData.confirmPassword)}
+          isLoading={isLoading}
+          loadingText="Registering"
+        >
+          Register
         </Button>
-        {activeStep === steps.length - 1 ? (
-          <Button
-            colorScheme="orange"
-            onClick={handleSubmit}
-            isDisabled={!isStepComplete(activeStep) || isLoading}
-            isLoading={isLoading}
-            loadingText="Submitting"
-          >
-            Submit
-          </Button>
-        ) : (
-          <Button colorScheme="orange" onClick={handleNext} isDisabled={!isStepComplete(activeStep)}>
-            Next
-          </Button>
-        )}
       </Flex>
     </Box>
   );
-};
-
-
+}
